requirements.txt
six==1.15.0

django-email-queue